View Doctor Detail and Add Schedule for Doctor
